+++ lib/sidekiq.rb
@@ -258,3 +258,4 @@ module Sidekiq
 end
 
 require "sidekiq/rails" if defined?(::Rails::Engine)
+require "sidekiq/systemd"